+++ pyobjc-framework-Automator/Lib/Automator/__init__.py
@@ -11,7 +11,7 @@ def _setup():
 
     import AppKit
     import objc
-    from . import _metadata
+    from . import _metadata, _Automator
 
     dir_func, getattr_func = objc.createFrameworkDirAndGetattr(
         name="Automator",
@@ -21,7 +21,10 @@ def _setup():
         ),
         globals_dict=globals(),
         inline_list=None,
-        parents=(AppKit,),
+        parents=(
+            _Automator,
+            AppKit,
+        ),
         metadict=_metadata.__dict__,
     )