+++ pyobjc-framework-ModelIO/PyObjCTest/test_mdlanimation.py
@@ -1,4 +1,4 @@
-from PyObjCTools.TestSupport import TestCase, min_sdk_level
+from PyObjCTools.TestSupport import TestCase, min_sdk_level, min_os_level
 import ModelIO
 from objc import simd
 
@@ -8,7 +8,7 @@ class TestMDLAnimation(TestCase):
     def testProtocols(self):
         self.assertProtocolExists("MDLJointAnimation")
 
-    @min_sdk_level("10.13")
+    @min_os_level("10.13")
     def test_methods(self):
         self.assertResultHasType(
             ModelIO.MDLAnimationBindComponent.geometryBindTransform,