+++ pyobjc-framework-WebKit/PyObjCTest/test_javascriptcore.py
@@ -5,11 +5,4 @@ import JavaScriptCore
 
 class TestCallableMetadata(TestCase):
     def test_callable_metadata_is_sane(self):
-        self.assertCallableMetadataIsSane(
-            JavaScriptCore,
-            exclude_attrs={
-                # Private classes on macOS 10.13
-                "ABCDContact_ABCDContact_",
-                "ABCDGroup_ABCDGroup_",
-            },
-        )
+        self.assertCallableMetadataIsSane(JavaScriptCore)